Check that mnesia schema was created on the node with the same name as current.
The problem arises when Kelly previously been launched using automatically generated node name such as kelly@127.0.0.1.
Mnesia schema is created with such name. If after a while the kellys node name is changed, then this situation is
noticed and an error shown. It's wise to name the node right at the very beginning.