Use different jsx & thrift paths
The problem is the thrift uses old jsx v0.9.0 version, that itself contains internal dependency named nicefloats.
Previously we don't need to use float presentation. Now we need to. It turned out that the nicefloats was unavailable.
So I forked both thrift and jsx and modified them both to use this specific patched jsx repository.